+++ core/tabs/applications-setup/design-tools/audacity.sh
@@ -2,9 +2,14 @@
 
 . ../../common-script.sh
 
+LINUTIL_UNINSTALL_SUPPORTED=1
+APP_FLATPAK_ID="org.audacityteam.Audacity"
+APP_UNINSTALL_PKGS="audacity"
+
+
 installAudacity() {
 	printf "%b\n" "${YELLOW}Installing Audacity...${RC}"
-	if ! command_exists audacity; then
+	if ! flatpak_app_installed org.audacityteam.Audacity && ! command_exists audacity; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
 				"$ESCALATION_TOOL" "$PACKAGER" install -y audacity
@@ -13,12 +18,15 @@ installAudacity() {
 				"$ESCALATION_TOOL" "$PACKAGER" -S --needed --noconfirm --cleanafter audacity
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.audacityteam.Audacity
-	            fi
-	            exit 1
+	        	printf "%b\n" "${YELLOW}No native package configured for ${PACKAGER}. Falling back to Flatpak...${RC}"
 	            ;;
 	    esac
+        if command_exists audacity; then
+            return 0
+        fi
+        if try_flatpak_install org.audacityteam.Audacity; then
+            return 0
+        fi
 	else
 		printf "%b\n" "${GREEN}Audacity is already installed.${RC}"
 	fi
@@ -26,6 +34,9 @@ installAudacity() {
 
 uninstallAudacity() {
 	printf "%b\n" "${YELLOW}Uninstalling Audacity...${RC}"
+	if uninstall_flatpak_if_installed org.audacityteam.Audacity; then
+	    return 0
+	fi
 	if command_exists audacity; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +46,7 @@ uninstallAudacity() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter audacity
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive org.audacityteam.Audacity
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
@@ -59,4 +70,10 @@ main() {
 
 checkEnv
 checkEscalationTool
+if [ "$LINUTIL_ACTION" = "uninstall" ]; then
+    uninstall_app "$APP_FLATPAK_ID" "$APP_UNINSTALL_PKGS"
+    exit 0
+fi
+
+
 main
