+++ core/tabs/applications-setup/design-tools/gimp.sh
@@ -2,9 +2,14 @@
 
 . ../../common-script.sh
 
+LINUTIL_UNINSTALL_SUPPORTED=1
+APP_FLATPAK_ID="org.gimp.GIMP"
+APP_UNINSTALL_PKGS="gimp"
+
+
 installGIMP() {
 	printf "%b\n" "${YELLOW}Installing GIMP...${RC}"
-	if ! command_exists gimp; then
+	if ! flatpak_app_installed org.gimp.GIMP && ! command_exists gimp; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
 			    "$ESCALATION_TOOL" "$PACKAGER" install -y gimp
@@ -13,12 +18,15 @@ installGIMP() {
 	        	"$AUR_HELPER" -S --needed --noconfirm --cleanafter gimp
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.gimp.GIMP
-	            fi
-	            exit 1
+	        	printf "%b\n" "${YELLOW}No native package configured for ${PACKAGER}. Falling back to Flatpak...${RC}"
 	            ;;
 	    esac
+        if command_exists gimp; then
+            return 0
+        fi
+        if try_flatpak_install org.gimp.GIMP; then
+            return 0
+        fi
 	else
 		printf "%b\n" "${GREEN}GIMP is already installed.${RC}"
 	fi
@@ -26,6 +34,9 @@ installGIMP() {
 
 uninstallGIMP() {
 	printf "%b\n" "${YELLOW}Uninstalling GIMP...${RC}"
+	if uninstall_flatpak_if_installed org.gimp.GIMP; then
+	    return 0
+	fi
 	if command_exists gimp; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +46,7 @@ uninstallGIMP() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter gimp
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive org.gimp.GIMP
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
@@ -59,4 +70,10 @@ main() {
 
 checkEnv
 checkEscalationTool
+if [ "$LINUTIL_ACTION" = "uninstall" ]; then
+    uninstall_app "$APP_FLATPAK_ID" "$APP_UNINSTALL_PKGS"
+    exit 0
+fi
+
+
 main
