+++ core/tabs/applications-setup/design-tools/inkscape.sh
@@ -2,9 +2,14 @@
 
 . ../../common-script.sh
 
+LINUTIL_UNINSTALL_SUPPORTED=1
+APP_FLATPAK_ID="org.inkscape.Inkscape"
+APP_UNINSTALL_PKGS="inkscape"
+
+
 installInkscape() {
 	printf "%b\n" "${YELLOW}Installing Inkscape...${RC}"
-	if ! command_exists inkscape; then
+	if ! flatpak_app_installed org.inkscape.Inkscape && ! command_exists inkscape; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
 			    "$ESCALATION_TOOL" "$PACKAGER" install -y inkscape
@@ -13,12 +18,15 @@ installInkscape() {
 			    "$AUR_HELPER" -S --needed --noconfirm --cleanafter inkscape
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.inkscape.Inkscape
-	            fi
-	            exit 1
+	        	printf "%b\n" "${YELLOW}No native package configured for ${PACKAGER}. Falling back to Flatpak...${RC}"
 	            ;;
 	    esac
+        if command_exists inkscape; then
+            return 0
+        fi
+        if try_flatpak_install org.inkscape.Inkscape; then
+            return 0
+        fi
 	else
 		printf "%b\n" "${GREEN}Inkscape is already installed.${RC}"
 	fi
@@ -26,6 +34,9 @@ installInkscape() {
 
 uninstallInkscape() {
 	printf "%b\n" "${YELLOW}Uninstalling Inkscape...${RC}"
+	if uninstall_flatpak_if_installed org.inkscape.Inkscape; then
+	    return 0
+	fi
 	if command_exists inkscape; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +46,7 @@ uninstallInkscape() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter inkscape
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive org.inkscape.Inkscape
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
@@ -59,4 +70,10 @@ main() {
 
 checkEnv
 checkEscalationTool
+if [ "$LINUTIL_ACTION" = "uninstall" ]; then
+    uninstall_app "$APP_FLATPAK_ID" "$APP_UNINSTALL_PKGS"
+    exit 0
+fi
+
+
 main
