+++ core/tabs/applications-setup/design-tools/openshot.sh
@@ -2,9 +2,14 @@
 
 . ../../common-script.sh
 
+LINUTIL_UNINSTALL_SUPPORTED=1
+APP_FLATPAK_ID="org.openshot.OpenShot"
+APP_UNINSTALL_PKGS="openshot-qt"
+
+
 installOpenShot() {
 	printf "%b\n" "${YELLOW}Installing OpenShot...${RC}"
-	if ! command_exists openshot; then
+	if ! flatpak_app_installed org.openshot.OpenShot && ! command_exists openshot; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
 			    "$ESCALATION_TOOL" "$PACKAGER" install -y openshot-qt
@@ -13,12 +18,15 @@ installOpenShot() {
 			    "$AUR_HELPER" -S --needed --noconfirm --cleanafter openshot
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.openshot.OpenShot
-	            fi
-	            exit 1
+	        	printf "%b\n" "${YELLOW}No native package configured for ${PACKAGER}. Falling back to Flatpak...${RC}"
 	            ;;
 	    esac
+        if command_exists openshot; then
+            return 0
+        fi
+        if try_flatpak_install org.openshot.OpenShot; then
+            return 0
+        fi
 	else
 		printf "%b\n" "${GREEN}OpenShot is already installed.${RC}"
 	fi
@@ -26,6 +34,9 @@ installOpenShot() {
 
 uninstallOpenShot() {
 	printf "%b\n" "${YELLOW}Uninstalling OpenShot...${RC}"
+	if uninstall_flatpak_if_installed org.openshot.OpenShot; then
+	    return 0
+	fi
 	if command_exists openshot; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +46,7 @@ uninstallOpenShot() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter openshot
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive org.openshot.OpenShot
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
@@ -59,4 +70,10 @@ main() {
 
 checkEnv
 checkEscalationTool
+if [ "$LINUTIL_ACTION" = "uninstall" ]; then
+    uninstall_app "$APP_FLATPAK_ID" "$APP_UNINSTALL_PKGS"
+    exit 0
+fi
+
+
 main
